requirements.txt
flet==0.70.0.dev5089
motor
pydantic
pytest
pytest-asyncio
python-dotenv
